global layout with menu and footer, adapt pages to global layout
