+++ pages/Credentials.vue
@@ -1,73 +1,36 @@
 <template>
-    <section class="_main bg-light row align-items-center justify-content-center justify-content-lg-start justify-content-md-center justify-content-sm-center p-0">
-        <div id="widget" class="_form d-black bg-w shadow-lg text-center">
-            <div class="_toggle-menu position-sticky d-flex justify-content-end col-12 align-items-center px-3">
-                <a id="toggle-menu" @click="menuTrigger">
-                    <div id="dash-1" class="_dash my-2"></div>
-                    <div id="dash-2" class="_dash my-2"></div>
-                    <div id="dash-3" class="_dash my-2"></div>
-                </a>
-            </div>
-            <div class="_content justify-content-center d-flex align-items-center ">
-                <div id="content">
-                    <h2 class="_animation-fade">{{$t('CREDENTIALS.TITLE')}}</h2>
-                    <div class="_search">
-                        <form action="" id="search-form">
-                            <input name="search-input" type="text" v-model="search" :placeholder="$t('CREDENTIALS.SEARCH')">
-                        </form>
-                    </div>
-                    <div class="_scrollable d-flex flex-column align-items-center">
-                        <NuxtLink class="_card d-flex" v-for="credential in filteredList" v-bind:key="credential.id" :to="'/Credential?id='+encodeURIComponent(credential.id)" >
-                            <div class="col-10 d-flex align-items-center">
-                                <div>
-                                    <h5 class="mb-1">{{$t('CREDENTIAL.TYPE.' + credential.type[credential.type.length-1])}}</h5>
-                                    <p class="text-truncate" style="max-width: 12em">by {{credential.issuerName ? credential.issuerName : credential.issuer}}</p>
-                                </div>
-                            </div>
-                            <div class="col d-flex justify-content-end">
-                                <i class="bi bi-bookmark-star-fill _icon-active"></i>
-                            </div>
-                        </NuxtLink>
-                    </div>
-                    <NuxtLink to="/AddCredential" class="btn btn-primary py-2" style="width: 70%;">
-                      <i class="bi bi-plus-lg me-2"></i>Request Credential
-                    </NuxtLink>
-                </div>
-                <div id="menu-content" class="_menu-content hide">
-                    <ul>
-                        <li>
-                            <NuxtLink to="/credentials">{{$t('MENU.CREDENTIALS')}}</NuxtLink>
-                        </li>
-                        <li>
-                            <NuxtLink to="/connections">{{$t('MENU.CONNECTIONS')}}</NuxtLink>
-                        </li>
-                        <li>
-                            <NuxtLink to="/settings">{{$t('MENU.SETTINGS')}}</NuxtLink>
-                        </li>
-                        <li>
-                            <a @click="logout">{{$t('MENU.LOGOUT')}}</a>
-                        </li>
-                    </ul>
-                </div>
-            </div>
-            <div class="_copyright _blue-color d-flex align-items-center justify-content-center">
-                <a id="copyright" href="https://walt.id/" target="_blank">{{copyright}}</a>
-            </div>
+    <div>
+        <h2 class="_animation-fade">{{$t('CREDENTIALS.TITLE')}}</h2>
+        <div class="_search">
+            <form action="" id="search-form">
+                <input name="search-input" type="text" v-model="search" :placeholder="$t('CREDENTIALS.SEARCH')">
+            </form>
         </div>
-    </section>
+        <div class="_scrollable d-flex flex-column align-items-center">
+            <NuxtLink class="_card d-flex" v-for="credential in filteredList" v-bind:key="credential.id" :to="'/Credential?id='+encodeURIComponent(credential.id)" >
+                <div class="col-10 d-flex align-items-center">
+                    <div>
+                        <h5 class="mb-1">{{$t('CREDENTIAL.TYPE.' + credential.type[credential.type.length-1])}}</h5>
+                        <p class="text-truncate" style="max-width: 12em">by {{credential.issuerName ? credential.issuerName : credential.issuer}}</p>
+                    </div>
+                </div>
+                <div class="col d-flex justify-content-end">
+                    <i class="bi bi-bookmark-star-fill _icon-active"></i>
+                </div>
+            </NuxtLink>
+        </div>
+        <NuxtLink to="/AddCredential" class="btn btn-primary py-2" style="width: 70%;">
+            <i class="bi bi-plus-lg me-2"></i>Request Credential
+        </NuxtLink>
+    </div>
 </template>
 
 <script>
-
-import {menuTransitionShow, menuTransitionHide} from '../helpers/menuTransation'
-import {config} from '/config.js'
 
 export default {
   name: 'Credentials',
   data() {
     return {
-      copyright: config.copyright,
-      trigger: true,
       search: '',
       credentials: [
           {
@@ -155,21 +118,6 @@ export default {
     const credList = await $axios.$get("/api/wallet/credentials/list")
     const credentials = credList.list
     return { credentials }
-  },
-  methods:{
-    menuTrigger: function(){
-          if(this.trigger === true){
-              menuTransitionShow()
-              this.trigger = false
-          }
-          else{
-              menuTransitionHide()
-              this.trigger = true
-          }
-    },
-    logout: async function() {
-        await this.$auth.logout();
-      },
   }
 };
 </script>
