+++ pages/Login.vue
@@ -1,8 +1,5 @@
 <template>
   <div>
-    <Notice></Notice>
-    <section class="_main bg-light row align-items-center justify-content-center justify-content-lg-start justify-content-md-center justify-content-sm-center">
-    <div id="widget" class="_form d-grid align-items-center bg-white shadow-lg text-center">
       <div :class="this.isSignin ? '_fade' : 'hide'">
         <img :src="this.logo.path" width="50px" :alt="this.logo.alt" />
         
@@ -47,7 +44,6 @@
             <img :src="locale.flag" width="20px" height="15px" :alt="locale.iso" class="me-2 border border-white" />
           </a>
         </form>
-        <a id="copyright" class="_animation-fade" href="https://walt.id/" target="_blank">{{copyright}}</a>
       </div>
       <div :class="this.isSignup ? '_fadehi' : 'hide'">
           <h2>Sign up</h2>
@@ -84,7 +80,6 @@
               </div>
               
             </form>
-          <a id="copyright" href="https://walt.id/" target="_blank">{{copyright}}</a>
       </div>
       <div :class="this.isResetPassword ? '_fadehi' : 'hide'">
           <h2>Reset password</h2>
@@ -100,23 +95,18 @@
                   <a @click="toSignIn">Already know your account? Login</a>
               </div>
             </form>
-          <a id="copyright" href="https://walt.id/" target="_blank">{{copyright}}</a>
       </div>
-    </div>
-    </section>
   </div>
 </template>
 
 <script>
 import ErrorMessage from '@/components/ErrorMessage.vue'
-import Notice from '@/components/Notice.vue'
 import {config} from '/config.js'
 
 export default {
   name: 'Login',
   components:{
     ErrorMessage,
-    Notice
   },
   data () {
     return {
